change to interpolate and refactor
