+++ torchxrayvision/baseline_models/jfhealthcare/__init__.py
@@ -9,6 +9,7 @@ import json
 import pathlib
 import torch
 import torch.nn as nn
+from ... import utils
 
 
 class DenseNet(nn.Module):
@@ -76,13 +77,14 @@ class DenseNet(nn.Module):
             raise (e)
 
         self.model = model
-        self.upsample = nn.Upsample(size=(512, 512), mode='bilinear', align_corners=False)
 
         self.pathologies = self.targets
 
     def forward(self, x):
         x = x.repeat(1, 3, 1, 1)
-        x = self.upsample(x)
+        
+        x = utils.fix_resolution(x, 512, self)
+        utils.warn_normalization(x)
 
         # expecting values between [-1024,1024]
         x = x / 512
