+++ torchxrayvision/baseline_models/xinario/__init__.py
@@ -6,6 +6,7 @@ import torch.nn as nn
 import torchvision
 import pathlib
 import torchxrayvision as xrv
+from ... import utils
 
 
 class ViewModel(nn.Module):
@@ -63,12 +64,6 @@ class ViewModel(nn.Module):
             print("Loading failure. Check weights file:", self.weights_filename_local)
             raise e
 
-        self.upsample = nn.Upsample(
-            size=(224, 224),
-            mode='bilinear',
-            align_corners=False,
-        )
-
         self.norm = torchvision.transforms.Normalize(
             [0.485, 0.456, 0.406],
             [0.229, 0.224, 0.225],
@@ -76,7 +71,9 @@ class ViewModel(nn.Module):
 
     def forward(self, x):
         x = x.repeat(1, 3, 1, 1)
-        x = self.upsample(x)
+        
+        x = utils.fix_resolution(x, 224, self)
+        utils.warn_normalization(x)
 
         # expecting values between [-1024,1024]
         x = (x + 1024) / 2048
